Update Visual Source (Using PWCT) - language/src/vmgc.c - Better Code

+++ extensions/android/ringlibsdl/project/jni/ring/src/vmgc.c
@@ -209,7 +209,7 @@ RING_API void * ring_malloc ( size_t size )
     /* Check Memory */
     if ( pMemory == NULL ) {
         printf( RING_OOM ) ;
-        printf( "Ring (MALLOC) - Can't allocate %d bytes\n",size ) ;
+        printf( "Ring (MALLOC) - Can't allocate %lu bytes\n",(unsigned long) size ) ;
         exit(0);
     }
     return pMemory ;
@@ -227,7 +227,7 @@ RING_API void * ring_calloc ( size_t nitems, size_t size )
     /* Check Memory */
     if ( pMemory == NULL ) {
         printf( RING_OOM ) ;
-        printf( "Ring (CALLOC) - Can't allocate %d bytes\n",nitems*size ) ;
+        printf( "Ring (CALLOC) - Can't allocate %lu bytes\n",(unsigned long) (nitems*size) ) ;
         exit(0);
     }
     return pMemory ;
@@ -240,7 +240,7 @@ RING_API void * ring_realloc ( void *ptr, size_t size )
     /* Check Memory */
     if ( pMemory == NULL ) {
         printf( RING_OOM ) ;
-        printf( "Ring (REALLOC) - Can't allocate %d bytes\n",size ) ;
+        printf( "Ring (REALLOC) - Can't allocate %lu bytes\n",(unsigned long) size ) ;
         exit(0);
     }
     return pMemory ;
